clean:considered the inputLines method as part of the MemoManager class responsibility
